Pulled processors and validators into their own files
