+++ package.json
@@ -31,6 +31,7 @@
   "dependencies": {
     "@fastify/cors": "11.0.1",
     "@fastify/http-proxy": "11.1.2",
+    "@tryghost/errors": "^1.3.7",
     "dotenv": "16.5.0",
     "fastify": "5.3.0",
     "pino-pretty": "10.3.1",
